Fixing odd random behaviour
Setting a new random seed for the pseudo-random generator to fix the problem of getting the same random numbers every iteration which results in odd behaviour if you change the convar ttt2_deathgrip_chance below the default 0.5. Before this change the deathgrip would still occur every second round even if you did set the ConVar to 0.1.